Implement spatial buy-to-let behaviour
Investors now decide whether to buy or not a new investment property and where based on computing a probability to buy a new property in each region, which is proportional to the expected yield of a newly bought property in that region. A random number is then drawn for deciding whether to buy or not and in which region.