Break implementDeaths at Demographics into three runs through regions
Dividing the decision on who to kill, the removal from household's home region and inheritance in three separate steps or three consecutive runs through regions allows to: 1) avoid needing a copy of all the households for safe removal, and 2) solves a previous bug when the inheriting household is itself killed in the same time step having just changed region due to inheritance (that was leading to the household not being killed, keeping its home but forgetting about any housePayments).

This solves a bug leading to wrong counts of the different types of households due to them moving around.